Handle cloud media_collection_id changes
CloudMediaProviders expose a users media collection to the MediaProvider via
a media_collection_id

When a user has the PhotoPicker UI open, the MediaProvider needs to
avoid races where media from a different collection is shown in the
UI. We achieve this with 2 API modifications:

1. CMPs will attach the media_collection_id to the returned query
cursor. This allows the MP to verify that the results match the
current media collection.
2. CMPs will attach the media_collection_id to the MP when notifying it
about media change events. This allows the MP to prioritise a sync for
the notification correctly. For instance, if the UI is currently open
and the media_collection_id changes, MP can refresh the UI so that
users always see the latest data.

Also addressed API review feedback to throw a SecurityException if the
media changed notification failed which can only happen if the caller
is not the enabled cloud media provider.

Test: atest CloudPhotoPickerTest
Bug: 214029148
Change-Id: I8b34354e93475d60116c4e2dc6834b1ca6bbc1d5