Fix a11y focus traversal order for Photos|Albums and WorkProfile button
Accessibility talkback would go directly from privacy text to the first
item in recycler view (which is "Recents"). Fixed this behavior by
following the below order:
1. Privacy text
2. PhotoPicker Cross Button
2. Photos Tab
3. Albums Tab
4. Switch to work profile button
5. First item in recycler view (that is recents or any other date
   header)

Fixed the above for Albums Tab as well. This does not fix focus traversal
for Albums Photos Grid which is tracked in a separate bug.

Bug: 207301848
Bug: 207319188
Test: Manual (see screen recording on the bug)
Change-Id: I36ebea1ee602e00701bbe22fedf35aed6c451c41