Add content description for special format files in Photo grid
Content description would specify if the item is a Photo or Video or GIF
or Motion Photo along with the date taken of the item.

Example:
"Photo taken on Feb 22, 2022, 2:22 AM" or
"Video taken on Feb 2, 2022, 2:02 PM" or
"GIF taken on Feb 22, 2022, 2:22 AM" or
"Motion Photo taken on Feb 2, 2022, 2:02 PM"

We have added separate content descriptions for the above as the
thumbnails in the photo grid has indicators/icons for GIF and Motion
Photo and Video. So, to give the same accessibility to all users we have
reflected the same in the content description.

Bug: 207057209
Test: atest com.android.providers.media.photopicker.data.model.ItemTest#testGetContentDescription
Change-Id: Ibb0a4b6903d228b3f55e6d57f724a5d7461212ec