Allowing clone user to access public volumes
In MediaProvider, each volume have an associated owner. As clone profile
and owner profile is sharing same instance of MediaProvider, the public
volume are associated with owner user. Adding a field in MediaVolume to
identify if volume is public, for public volume allowing clone user to access
the volume in MediaProvider even if it is associated to owner user.

Bug: 213104675
Test: Manually tested by creating an sample app installed in clone app
trying to access vitual sdcard(public volume)

Change-Id: Ieafead22646384160687be57e58481522a0679f8