Update GENERATION_MODIFIED while updating SPECIAL_FORMAT column
Make sure to update GENERATION_MODIFIED column while updating any row.
This is essential for the Picker DB to know which rows were modified.
This is a part of the CloudMediaProvider Contract that the providers
must always update GENERATION_MODIFIED to reflect that a row  has been
modified.

db.update() does not update the GENERATION_MODIFIED, whereas
QueryBuilder takes care of it internally.

Bug: 210400433
Test: atest IdleServiceTest

Change-Id: If6d9bb7efb10581bc5c41552f4b6c11ba6c50a60