Change muteButton implementation to remove selector usage
Previously, muteButton used "selector" to change the drawable from
ic_volume_off->ic_volume_up based on the state of the ImageButton. This
caused issue with talkback announcements where it announced "selected"
for "mute" state. Changed the implementation to remove the use of
ImageButton state and programatically set the mute/unmute icons.

Bug: 223500623
Test: atest android.photopicker.cts.PhotoPickerTest#testMultiSelect_previewVideoMuteButtonInitial
Test: atest android.photopicker.cts.PhotoPickerTest#testMultiSelect_previewVideoMuteButtonOnSwipe
Change-Id: Ibf5e1e3e4c1c978fba2c5846b65977f5315e4f3e