[a11y] Unblock items less visible due to profile button
Add extra padding at the bottom of the recyclerview so that profile
button does not obstruct visibility of photos or albums.

The extra padding for profile button visibility is only added if profile
button maybe visible in that page (for example: Photos and Albums Tab if
work profile is present). However, the padding will not be present if
the profile button will not show (for example: if work profile is not
present or in Album Photos tab).

We do not add extra padding for bottombar in case of selection is made
in multi select mode. This is because the padding is already present
due to virtue of work profile button.

Bug: 223710116
Test: manual (screenshot attached to the bug)
Change-Id: I83e46bf590153d0792a775b250e5ad6c3c79d449