Add support to get file attributes through xattr
Adding the code to fetch file metadata stored in the xattributes. The
SQLite call will only be used as a fallback in case the xattr call fails. This
also includes the code to serialize/deserialize the FileAccessAttributes
class to and from byte[]. Approx. 50% reduction in time taken for the
file open call is expected after we migrate the file open calls to
xattrs.

Bug: 227354088
Test: Tested on device by creating a random text file, setting xattrs
corresponding to the file and then reading the file.

Change-Id: I50d357d21686ba696b76fd83b2177b77d27b901a