Add external volume names to PhotoPicker media collection id
To support unmounting volumes and hiding the content from the picker,
we embed the mounted external volumes as part of the media collection
id. That way, during a picker sync if the mount state has changed, we
force a new sync with the latest mount state.

To avoid trashing when mounting multiple volumes at boot, we remove
the picker syncs at volume mount. These aren't required anymore since
we force a sync at picker launch and the syncs are quite fast.

Test: atest ExternalDbFacadeTest
Bug: 228312259
Change-Id: I0dba349474f36d32a82174f81287df260c395f56