Skip id reuse fix for internal db
To avoid id resue in MP databses, a fix was made in ag/16603120 to start
over id mappings from a billion and backup next row id in external
storage, which can be used in case of DB rollback. However, it was
identified in b/229153534 that the ringtones manager can get affected
by ids change. Hence this change skips id reuse fix for internal db.

Test: atest StableUrisTest
Bug: 216743720
Change-Id: I82c98e579063fe377c9f9cc83bbbc76f54d3e4e7