Load picker URIs with PhotoPicker Glide module
Picker URIs of the form content://<authority>/media should be loaded
with the PhotoPicker custom glide module. This module ensures that
Glide URI data requests send the size as an argument to the backing
provider.

This ensures that cloud media providers get a thumbnail request as a
call to openPreview instead of openMedia which requires fetching the
entire original media from the network. For videos, openMedia can be
very slow and lag the thumbnails loading on the PhotoPicker grid.

Test: Manual
Bug: 230024255
Change-Id: Ia4fa5cce28b11b6a3e5b929bb1ad8afe03c6e4ff