Fix openPreview and openMedia handling
Previously, we didn't pass the EXTRA_PREVIEW_THUMBNAIL flag to cloud
providers as part of openPreview to indicate when they should return a
thumbnail or video preview media.

Previously, we called openPreview as part of the ContentProvider
openTypedAsset calls and openMedia as part of the openFile
calls. This worked fine but a different policy was needed to use
ExoPlayer's ContentDataSource internal implementation for video
preview. Since ContentDataSource always called openTypedAsset, we now
only call openPreview as part of openTypedAsset.

Test: Manual
Bug: 230725491
Change-Id: I8a18a53023300cb879bca2ef20b72143aa169b0c