Show toast when cloud provider is first enabled
This is a workaround to ease app developer testing pending the final
UX when the cloud provider launches.

We persist state indicating a cloud provider has just been enabled and
show a toast on next picker launch if the state is still true. The
state is cleared if either of the following occur:
1. Cloud provider is unset
2. Cloud provider is uninstalled
3. Toast has already be shown

Test: Manual
Bug: 228458085
Change-Id: Ifc75fb6b6e92c59bbb8341dab1cbede0a3d13199