MediaProvider to report if FUSE passthrough is enabled
Relying on the persist.sys.fuse.passthrough.enable system property to
check if FUSE passthrough is enabled is not a reliable solution, as
shown by mixed builds with GSI. The problem arises when an underlying
system not supporting FUSE passthrough (e.g., Android R) runs with an
Android S+ GSI which wants to use FUSE passthrough, thus has
persist.sys.fuse.passthrough.enable set to true.
Android works fine as MediaProvider (FuseDaemon) catches the exception
that the kernel does not support FUSE passthrough, but the Android S+
tests assume that FUSE passthrough is enabled because of the system
property.

This change extends the MediaProvider hidden API with a call to test if the FUSE
passthrough feature is being used or not by the FUSE daemon.

Bug: 207733510
Test: android.scopedstorage.cts.device.ScopedStorageDeviceTest
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: Ie31be07d97a1f9ad4402f7f44490c924e883a0a9