Render remote video preview in picker UI process
Previously, the local provider (PhotoPickerProvider) directly created
a RemoteSurfaceController for video preview. This meant that an
ExoPlayer instance was created in the main MediaProvider process.

This worked on Android T but fails on R&S because MediaProvider
doesn't have the correct selinux policies to play video:
aosp/Ie5acb77b2f446ee8af6cf384fd5a66bf64a15752

Now, we created a RemoteVideoPreviewProvider which is a lean
CloudMediaProvider that only serves to create a
RemoteSurfaceController in the picker UI process. The controller
created here is marshalled back to the PhotoPickerProvider which is
the local provider in the MediaProvider process

Test: atest PhotoPickerTest on S device passes
Bug: 233848463
Bug: 232498432

Change-Id: Ib88713524f928af3cac27072e131b64a65408129