Disable special format detection for existing files
Temporary quick fix by disabling the culprit code which updates a huge
number of image rows and also updates their generation_modified. This
indicates other providers that huge number of rows have been updated and
hence may lead to huge computation and memory extensive processes to
trigger.
To avoid this, we will not be updating the generation_modified for
special_format detection as this is a hidden column added only for
Picker. We will enable special format detection on existing files once
the fix to not update generation_modified is merged.

Bug: 235558753
Bug: 235426236
Bug: 236620024
Test: presubmit
Change-Id: I1d6e528b4e614c367be30cd65f9e5147d756921a