Support PhotoPicker explicitly launched from DocumentsUi
PhotoPicker can be explicitly opened from DocumentsUi, PhotoPicker
should be launched even if the intent requests non-media mimetype
filters.

In the case that the intent contains non-media mimetypes, PhotoPicker
shows all photos and videos.

Previously, we were checking if the intent was forwarded. This has a
strong assumption that *only* DocumentsUi sends a forwarded intent to
PhotoPicker. But it was observed that some apps can launch GET_CONTENT
using ACTION_CHOOSER, in this case PhotoPicker receives a forwarded
intent from ChooserActivity. Update code to *only* launch PhotoPicker if
the intent was referred by DocumentsUi.

Bug: 231413552
Test: manually
Change-Id: If6031c70b4b6db0695f13901df2fe8ed2c967036