Avoid posting unnecessary transcoding notification tasks
Return early for files that do not support transcoding (like non-video
files), this helps in reducing the number of tasks posted for the
BackgroundThread to execute.

This helps us avoid memory issues if there is a bulk file open and the
BackgroundThread is busy.

Also, return early in TranscodeHelper#onFileOpen to avoid database query
for a file which does not support transcoding.

Bug: 235558753
Test: atest TranscodeHelperTest
Change-Id: Ib47277e138897113aa309224aecf3da05028a4f3