Cache the root inodes with backing fd
With the previous logic we always set entry timeout to
0 when we looking up for inode with bpf and
backing_fd. We're doing this because this part of FS
is backed and will revalidate, lookup (and do any other
operations) without touching the userspace.

But lookup execution is happening on parent and if
we set entry timeout to 0 for `data` folder we're
expecting to receive lookup(Android, data) to the
userspace all the time (this is something we're able
to see now).

On this change I'm making sure that entry cache is 0
only for cases where we inside the backing fs.

Why couldn't we do it before?

I remember Zim mentioned that after setting entry
cache to a large number for a `data` inode we starting
receiving another unwanted userspace operations. I was
able to reproduce it as well.

Problem: go/fuse-loosing-inode-with-backing

But now seams like this problem is fixed on the Kernel
and libfuse sides, and we can finally execute the proper fix
(cache `data|obb` entries) on fuse daemon's side as well.

Checked it locally, so we stopped receiving redundant
lookups after this change and finally not affecting any
other backing and bpf related logic.

Bug: 219958836
Test: Manually checked that we’re not receiving redundant
lookups anymore.
Test: atest --test-mapping packages/providers/MediaProvider
Signed-off-by: Dmitrii Merkurev <dimorinny@google.com>
Change-Id: I82b6efb82bd8032ab37d241806b446b88fc954a2