Make sure we’re not caching readdirplus lookup for backed files
Fuse-bpf driver implementation doesn’t support providing backing_fd and bpf program
as a part of readdirplus lookup. So we make sure here we're not making any lookups on backed files because we want to receive separate lookup calls for them later to set backing_fd and bpf.

Bug: 219958836
Test: Booted device 5 times to make sure we’re not receiving redundant
lookups anymore.
Test: atest --test-mapping packages/providers/MediaProvider
Change-Id: I579cdf965088651be84b70fcbc25ed3b4efaecee
Signed-off-by: Dmitrii Merkurev <dimorinny@google.com>