Don't redact length bytes for gps data in ilst box
Some video files can have gps data stored in XYZ boxes of item list(or
ilst) box. For e.g.,
```
Atom ilst @ 2221 of size: 88, ends @ 2309
    ...
    Atom ©xyz @ 2259 of size: 50, ends @ 2309
        Atom data @ 2267 of size: 42, ends @ 2309
```
Previously, we redacted size bytes of these boxes which may make some of
the parsing logic to assume size of the file as the size of this box.
To fix this isssue, we now redact only the header tags and data values
while keeping the size bytes intact. E.g.,
```
Atom ilst @ 2221 of size: 88, ends @ 2309
    ...
    Atom  @ 2259 of size: 50, ends @ 2309
         Atom  @ 2267 of size: 42, ends @ 2309
```

Bug: 220241280
Test: atest com.android.providers.media.util.IsoInterfaceTest

Change-Id: Id017e819d9fcedfe16777f9bc426ef5b2f502886