Add explanation for W_E_S permission check for non-legacy Installers
WRITE_EXTERNAL_STORAGE (W_E_S) permission should be a no-op for
non-legacy apps. But it was still being used by Installer apps as a
necessary permission. Since this check is in frameworks code, which
cannot be updated outside dessert releases, we decided to have
consistent checks in MediaProvider code as well.

Another reason to not add non-legacy check for W_E_S in MediaProvider is
because for Android S+, we always call into the frameworks API which
does not check the targetsdk.

Bug: 242862069
Test: atest ScopedStorageHostTest#testCheckInstallerAppAccessToObbDirs
Change-Id: I6f0ffc7a750a2bfe5c53ca8010094e03a782c700