Do not throw exception while scanning all volume paths
Currently getAllowedVolumePaths throws an exception if no
volume paths exist in the volume cache for any user. This can
cause failures even when the clients are looking for volume paths in
other users. This happens today when volumes corresponding to the
clone profile are not correclty loaded by the volume cache. This fix is
meant to unblock such cases.

Bug: 242009547
Test: Not able to reproduce the bug, relying on presubmit tests
Change-Id: I0c9410192d7b096cc3304a46afbdbb49f7234c4e