Move test apks from `java_resources` to `data`
Test apks should be accessed by Tradefed at runtime. But the test apks
declared in `java_resources` will be packaged into the output Jar/apk
file, and Tradefed can't access the apks packaged in other files.

This test works, since current testing infrastructure automatically
detects and builds the test apks. But we want to disable that feature,
because it adds and builds some unneccessary dependencies. So that,
this change moves those test apks to `data` property to fix the issue.

Test: atest MediaProviderTests
Change-Id: I9c347d9635d50ab0584ba63d0ec23d6df2de01f7