Introduce CloudProviderUtils and CloudProviderInfo
Move CloudProviderInfo class from PickerSyncController into its own
.java file and make it public.

Move utility methods for retrieving a list of CloudProviderInfo-s for
available MediaCloudProvider-s (getSupportedCloudProviders()) from
PickerSyncController to the (new) CloudProviderUtils.

Both CloudProviderInfo and CloudProviderUtils are now accessible by the
code running in the UI (:PhotoPicker) process, which is ultimately the
goal of this change.

Bug: 202141897
Bug: 195009187
Test: m MediaProvider
Test: m com.google.android.mediaprovider
Test: atest CtsPhotoPickerTest
Test: atest MediaProviderTests
Change-Id: Ic6e44e7d3c9c4758bee181927a3c6d5bcf50a03e