Add soft support for height/width/orientation columns for picker uris
Querying picker uris for unsupported columns throws an
IllegalArgumentException. Since we will soon be adding support to query
height/width/orientation columns soon, so we don't want to throw
Exceptions for such an app. We are returning a null value as default,
but the support for actual value will be added soon.

This is related to app compatibility for GET_CONTENT takeover. Many
DocumentsProviders in DocumentsUI do not throw error for unsupported
columns. This is not part of the API contract. This leads to
uncertain behavior and assumptions on app side. We will continue to
throw error for unsupported columns except for heigh/width/orientation
as we think those are valid columns that would be supported soon.

Bug: 247744230
Test: atest PickerDbFacadeTest
Change-Id: Ic054d73ee9407b46fe85f297e30f7dc7db822a44