Add migration code piece to migrate for internal
Added code to migrate data for internal db. Added a new job which will
be attempted to run once in 7 days on device idle and charging state
and backs up data to leveldb. Also added safety checks to avoid null
pointer exception for leveldb operations in FUSE layer if leveldb
setup is not successful. Added logic in device config listener to
setup leveldb connection if the flag is enabled.

Test: atest StableUriIdleMaintenanceServiceTest
Bug: 239415085

Change-Id: I0ab005d9dbdc77cf81604b413213e34b02f62a71