Fix database queries for Favorites album
Previously cloud + local items that have been marked as favorite in
cloud media app did not show in the Favorites album. The cover of
favorites album was also missing and the count of items in the album was
incorrect.

With this change, the above issues in the Favorites album will be fixed
by correcting the picker db query to fetch favorited items. All items
marked as favorited in cloud provider app or in external.db will be
deduped and merged in the same album.

Bug: 238730919
Test: Manually tested for all scenarios
Change-Id: I39bbef9cdc64967a2e737958cf2adc8e19910fe0