Fix orientation issue in PhotoPicker thumbnails
PhotoPicker shows some of the images with wrong orientation.
ThumbanailUtils#createImageThumbnail may return thumbnail with wrong
orientation because embedded JPEGs in the images may not have EXIF
markers.

To resolve this issue, we try to fetch the orientation using Glide
library APIs and send the ExifOrientationStream with extracted
orientation information.

Bug: 241198072
Test: Manual. Verified that raw image thumbnails are displayed with
right orientation. Screenshots attached in the bug.

Change-Id: I8ec4f862fdc55e08faf93d3cd843579c41ec0d59