Refactor access check SQL statements
Moved all access check related SQL query generation to a new method.
This simplifies SQLiteQueryBuilder a little and allows us to modify the
SQL query for new permission checks.

With these changes, access checks are split into following steps-
1) Check checkCallingPermissionGlobal -> If the app has global access
either due to MANAGE_EXTERNAL_STORAGE or uriGrants, we don't need any
where clause to restrict the query
2) hasAccessToCollection -> Check if app has corresponding read or
write access to the given collection. E.g., an app may have read
access to image collection if it has READ_MEDIA_IMAGES permission. In
this case we need not add any where clause to restrict the query.
3) If app has no permission or only limited access not a wider access
to all the items in the collection, we get the corresponding where
clause from getWhereForConstrainedAccess. This where clause will be
appended to the database query so that the given operation will only
be executed if the where clause allows the row to be visible.

Changed getWhereClauseForMatchableVisibleFromFilePath to use new methods
to improve the readability of the code. This removes
isCallingPackageLegacyWrite check and this check will be covered in
hasAccessToCollection -> checkCallingPermission[Audio|Video|Images]
check.

Removed where clauses like
```
matchSharedPackagesClause + " AND media_type=0 AND mime_type LIKE
'video/%'"
```
that allowed access to owned hidden media files. This is not necessary
as we already allow access to owned files which is a superset of owned
hidden files.

Bug: 259059840
Test: presubmit
Test: atest com.android.providers.media.AccessCheckHelperTest
Change-Id: I9fc3eee0de982ce98a5734d4a3ae4ec525ecc601