Fix flaky test
The test was flaky if it was run more than once. This happened because
the second test run used the same table where _ID would have been
updated due to Sqlite AUTO_INCREMENT. Changed the test to insert _ID
instead of using AUTO_INCREMENT so that we can predict the _ID value in
the test.

Bug: 240394046
Test: atest
com.android.providers.media.photopicker.UnreliableVolumeFacadeTest#queryMediaId

Change-Id: I7f528a9d5b78ea88dbae07462133fac46178a1c6