Create owner id to owner pkg name relation
Added code for creating owner id to owner package name relation and vice
versa in the backup. We set owner package id instead of owner package
name in the backup. When an application is uninstalled, all media rows
corresponding to it will be orphaned and would have owner package name
as null. This should not change if application is installed again.
Therefore, we are storing owner id instead of owner package name. On
package uninstallation, we delete the owner id relation from the backup.
All rows recovered for orphaned owner ids will have package name as
null. Since we also need to support cloned apps, we are storing a
combination of owner package name and user id to uniquely identify a
package.

Test: n/a
Bug: 259258592
Change-Id: I4ecda0fb5fadf0f8ec12b37482f693a9e0711c68