Unblock Settings UI while picker performs cloud media sync
Currently when user updates the cloud media provider from Settings
page, the sync starts right after the change in provider. However, the
sync blocks the UI thread instead of processing asynchronously.

With this change, the Media Provider thread will update the cloud
provider when requested, start a sync in the background thread and
return back the result to the UI process without waiting for the sync to
complete. This will make sure that the UI is not blocked while the cloud
media sync takes place.

Bug: 264999862
Test: Manual - details in the bug
Change-Id: I9f8e6d5ea87bc527ffd78757d6d5d501c4d59553