Use package names instead of authorities in CMP configs
Change the "format" of the Cloud Media Providers allowlist: list package
names instead of authorities.
For backward compatibility for any item in the allowlist that ends with
".couldprovider", we will are removing that suffix.

Deprecate config_default_cloud_provider_authority and introduce
config_defaultCloudMediaProviderPackage instead, for specifying the
"default system CMP".

Bug: 268174159
Test: atest MediaProviderTests
Change-Id: I6b35635bea31b7cc7ca0e094a7b59ec83917946b