Change PhotoPickerSettingsActivity's launch mode
Update PhotoPickerSettingsActivity's launch mode to single instance
so that users don't see multiple instances of the Settings page. This
would also avoid the problem of keeping the data in multiple instances
of the Settings page in sync.

This change also includes the Settings activity in recents. Since the
launch mode has changed to singleInstance, Settings activity will be a
part of a new task and the new task will only show up in recents with
this config change.

Bug: 267429483
Test: Manually tested
Change-Id: I51970647900170b1c327ded4b1d53e63164725a2