Make Picker Settings Activity window opaque
As the Settings activity's window was marked as translucent, when the
Settings activity was launched from the Picker activity, it did not
stop the Picker activity. As a result when the user navigated back to
the Picker activity, it's onRestart() method was not called and the
Picker activity did not reset as expected.

To solve this, we're marking Settings activity's window as NOT
translucent.

Bug: 266053476
Test: Verified on test device
Change-Id: I68c05018f506d2678a4fbf75e99133efd23b4f90