Introduce more Cloud-Media-in-Photo-Picker feature configs
Introduce two more configs for the Cloud-Media-in-Photo-Picker:
1) boolean ConfigStore.isCloudMediaInPhotoPickerEnabled()
2) boolean ConfigStore.shouldEnforceCloudProviderAllowlist()

The former is the "feature flag" for the feature.

The latter will allow us to by-pass the allowlist. We do not expect to
use this in "production", but this would help with testing and
droidfooding.

Bug: 268174159
Test: atest MediaProviderTests
Merged-In: I9c198faf832f3b1c9e9c55769bb8488b8b5245e0
Merged-In: Ia1ff25bea936fe24317269b3252853e962fd904a
Change-Id: I9c198faf832f3b1c9e9c55769bb8488b8b5245e0