Check recycler view position on banner insert
Check if the first completely visible item position in the recycler view
is the one just below the banner before scrolling to the banner position
on banner insertions in the photo picker.

During onViewCreated, before restoring the state, the first visible item
position is -1, and and we should not scroll to the banner position in
this case, else the previously saved recycler view position may get
overriden.

Test: Manual
Bug: 275337639
Change-Id: Ibe7b50c46a8a1aa394f2f4234ea020d40020f93f