Using userId received in onOpChange operation for media_grants reset.
Currently, there is no way for MediaProvider to know that an
opChange callback originated from root user or cloned useer app.
Consequently, media_grants for the package are removed for both
cloned an root user packages.
With ag/22934036, the capability to convey user space for app ops
was added to AppOpsManager.
This change allows MediaProvider to make use of the above, and
reset media_grants only for the app selected by the user.

Test: atest MediaGrantsAppOpStateTest
Bug: 278843665
Change-Id: If7b8185fe27284e8b265b2f535c62863155dfdce