Removing created_default_folders_<volId> entries during User removal.
When a cloned user is removed, `created_default_folders_<cloneVolId>`
type entries are not cleared from SharedPreferences (user 0).
As a result, if a new cloned user is created and gets the same id,
as the prevously removed one, default directories are not created,
due to persistence of `created_default_folders_<cloneVolId>`.

To fix this, we add a step to remove `created_default_folders_<volId>`
after a cloned user is removed.

Test: atest AppCloningMediaProviderHostTest
Bug: 267168181
Change-Id: I0d79ef258de71f39447c79fe4cf6e46593f66301