Skip cross-user install sessions call
We rely on the onIdleMaintenance call to ensure we remove all stale data
in MediaProvider. Given that we do not redirect PACKAGE_REMOVED_FULLY
broadcase for cloned packages, this call is the only source of stale
data deletion for stale entries corresponding to the cloned packages.
However, the getPackageInstaller().getAllSessions() call used to check
if a package corresponding to a media provider entry is removed,
requires INTERACT_ACROSS_USERS_FULL permission. This change ensures
this check is skipped for cloned users since the cloned packages are
installed by just copying over the packages from the primary profile and
don't really create install sessions. We also ensure that this check is
skipped for any other cases where getPackageInstaller().getAllSessions()
could lead to a cross-user call logging corresponding errors.

Test: Tested on device by skipping the checks in
MediaService.onScanVolume() to prune the file.
Bug: 245050056

Change-Id: Ibe5a9841631316caeee32c17b479e9ae98373f45