Add init method calls for triggering syncs from the UI
After this change, the sync trigger and query calls from the UI will be
decoupled. This is required because now the UI will query items in
pages. For each page query, we don't want to trigger a sync.

So, before UI start querying for items to display on the Picker UI,
an init call will be made which will be the trigger for Picker to start
syncing data with the respective providers.

Bug: 285890640
Test: Manual - more tests to follow
Change-Id: Ibf01a8cb63949df441f5121b55202340428ad1ac