Mock updateQuotaTypeForUri method
This method updates quota type in a background thread after a file
insert/update. We observe crash reports in MP tests when this method
tries to query database in the background thread after the testsetup
is removed. Mocking this call as this is not relevant for verifying any
test scenario.

Test: atest MediaProviderTests
Bug: 288345578
Change-Id: I7061710532e267f92ab5d66b8a073c1d20b8a27f