Improve media keep rules
Remove keep rules for providers.
Providers that are listed in a manifest get keep rules automatically
generated by aapt.
Providers that are not listed in a manifest are unreachable dead code
and shouldn't be kept.
Replace keep rules with annotations where applicable.

This change is a cleanup and a functional no-op.

Bug: 248580093
Change-Id: Iecfd561e9ede918dcf44e57d2ea60084ed2f01de