Use UserId of MediaProvider for package_user_id for adb shell
add_media_grants shell command allows us to add media_grants through
shell command. We don't have a good way to know the userId of the
corresponding package and userId of the shell may not be correct for
secondary users. Hence, we use userId of the MediaProvider which is more
accurate.

Bug: 286170881
Bug: 286171131
Bug: 289267744
Bug: 290036261
Bug: 289058662
Bug: 289059373
Test: atest android.scopedstorage.cts.device.StorageOtherFilesTest#other_readVisualMediaFiles
Change-Id: I7e3f0f4f78817c8ed784baaa02f773ac48746859