fix(PickerDbFacade): Fix missing album media in race conditions
Occasionally when a cloud sync is running, files synced to album_media
may be unavailable to the preloader / external apps that are trying to
query uris supplied by the picker.

Therefore, when we know there is a cloud sync running, write all synced
album_media to the media table as well so we can guarantee that the file
will exist if/when it is queried before the cloud sync finishes.

Eventually, the cloud sync will encounter this row, and just perform an
upsert over top of it bringing everything back to equilibrium.

Bug: b/294973487
Test: atest PickerDbFacade#testAddCloudAlbumMediaWhileCloudSyncIsRunning
Change-Id: I1026545b3c754729bf3777e58cf018e2e6e8e969