Revert the change to remove cloud query enabled checks for album media
Cloud queries can be disabled in Photo Picker by setting
PickerDbFacade.mCloudProvider to null. Currently, cloud queries are
disabled when a it is detected that a full sync is required with the
cloud provider and the existing media in the Picker DB is stale. This
also affects the cloud album content queries.

Ideally album content syncs and queries should be independent of media
syncs. However, removing the null check on PickerDbFacade.mCloudProvider
does not help us enable album queries independent of media syncs because
the queries use the value set in PickerDbFacade.mCloudProvider to
construct media URIs. Bug 294974183 tracks the long term fix for this.

Bug: 294974183
Test: atest PickerSyncControllerTest
Test: atest PickerDbFacadeTest
Change-Id: If038b66ab8b952f4d0ca1342dea4a0cdacdf161e