Use fixed ThreadPool in SelectedMediaPreloader for choosing multiple files to be open successfully
Assigning fixed number of threads irrespective of number of requests
coming to open Media file. Previously in selectedMediaPreloader we were assigning new thread to the execution of any file at runtime if needed. So to get rid of Timed out exception and to limit number of threads requesting to open media , changed this to fix thread pooling.

Bug: b/295809545
Test: In case of allow mutiple if you choose any number of media (less than 100) to add. All files will open successfully without FileNotFound Exception.
Change-Id: I9d5e328b7b487d2913b3a5c7044a8eadcc457eff