Update choose, change account banner action button
Before this change, the choose account and change account action buttons
were hidden from the picker banners.

With this change, these buttons will be visible iff the cloud provider
provides us with a non null intent to launch the account change activity
in the getMediaCollectionInfo API response bundle with a timeout of 100
ms. On click, these buttons would launch the activity using the intent
provided by the cloud provider.

Test: Manual
Bug: 274426228
Change-Id: I337919999d1450b4852802d559b7c6ac1fd0d217