Check WorkInfo.State is a sync takes longer than usual to complete
It's possible that WorkManager fails to run a sync after accepting the
sync request. Previously, since DataLayer does not  check work state from
work manager, in such scenarios query response would be stuck on sync futures
that did not clear.

With this change, if a sync work takes unusually long to run, DataLayer
would check the state of the task in WorkManager and unblock the query
if work is not pending.

Bug: 294023459
Test: atest PickerDataLayerTest
test: atest PickerSyncManagerTest
Change-Id: I8e9bf68ecfb4652c95eb273570175506cd1f730f